move tests for TFKerasApplication to specific test suite for models extractors
