+++ pliers/tests/extractors/test_image_extractors.py
@@ -8,7 +8,6 @@ from pliers.extractors import (BrightnessExtractor,
                                SharpnessExtractor,
                                VibranceExtractor,
                                SaliencyExtractor,
-                               TensorFlowKerasApplicationExtractor,
                                FaceRecognitionFaceLandmarksExtractor,
                                FaceRecognitionFaceLocationsExtractor,
                                FaceRecognitionFaceEncodingsExtractor)
@@ -55,25 +54,6 @@ def test_saliency_extractor():
     assert np.isclose(ms, 0.99669953, 1e-5)
     sf = result['frac_high_saliency'][0]
     assert np.isclose(sf, 0.27461971, 1e-5)
-
-
-def test_tensorflow_keras_application_extractor():
-    imgs = [join(IMAGE_DIR, f) for f in ['apple.jpg', 'obama.jpg']]
-    imgs = [ImageStim(im, onset=4.2, duration=1) for im in imgs]
-    ext = TensorFlowKerasApplicationExtractor()
-    results = ext.transform(imgs)
-    df = merge_results(results, format='wide', extractor_names='multi')
-    assert df.shape == (2, 19)
-    true = 0.9737075
-    pred = df['TensorFlowKerasApplicationExtractor'].loc[0, 'Granny_Smith']
-    assert np.isclose(true, pred, 1e-05)
-    true = 0.64234024
-    pred = df['TensorFlowKerasApplicationExtractor'].loc[1, 'Windsor_tie']
-    assert np.isclose(true, pred, 1e-05)
-    assert 4.2 in df[('onset', np.nan)].values
-    assert 1 in df[('duration', np.nan)].values
-    with pytest.raises(ValueError):
-        TensorFlowKerasApplicationExtractor(architecture='foo')
 
 
 def test_face_recognition_landmarks_extractor():
